shifted specific file ops to version-specific notebook runner
